pachyderm_sdk: separate config parsing into its own module
